Map scrolling finished
Next patch should fix dialogues to work with it and to fix the
long-living bug where dialogue is rendered from NPCs thus renders behind
entities and fog of war by centralising dialogueGUI and storing the
dialogue seperately. It'll save creating dialogue GUIs too which lags
the game a bit due to the rendering involved.