Bored of orthographic
I'm a long believer that isometric viewports for RPGs are vastly
superior for more serious-style games

So I spent an hour and fixed the rendering engine to support iso and
started redoing the tiles. Mouse is a bit off and entities don't
understand the new grid but it's a start.